requirements.txt
pytchat
pandas 
kafka-python
datetime
numpy
transformers
vncorenlp
torch
pyspark
findspark
pyvi
streamlit
matplotlib
